Ensure prettier is called last
