+++ lib/withReact.js
@@ -54,8 +54,11 @@ module.exports = (base) => ({
       rules: { 'react/jsx-props-no-spreading': 'off' },
     },
     {
-      files: ['*.stories.tsx?'],
-      rules: { 'react/prop-types': 'off' },
+      files: ['*.tsx?'],
+      rules: {
+        'react/prop-types': 'off',
+        'react/require-default-props': 'off',
+      },
     },
   ],
 })
